Moved microservice state component
